Configuration file for goesrecv
To accommodate configurable publishers (and potentially multiple per
DSP block) this commit also changes how data is published. It used to
be an add-on capability on the queue type. With this change it is the
responsibility of the blocks themselves. This makes it easier to add a
publisher that publishes sub-sampled data, an FFT, or stats about the
demodulator and/or decoder.