Add optional stats monitor to goesrecv
This displays a summary of demodulator and decoder statistics if
goesrecv is running in verbose mode. The interval at which this is
displayed is configurable via a command line option.

This mode doesn't introduce statistics that aren't already published
through the stats publishers. Moreover, it uses the exact same streams
and parses the JSON like any other third party subscriber would.