Replace goesdec for goeslrit
The purpose of goesdec is no longer to decode and write LRIT files
after the decoder was split into its own library. The only purpose of
this tool is to take a packet stream and write LRIT files.

This new version can either read packets from a list of files (in case
packets were persisted to disk) or subscribe to a packet publisher
directly (a goesrecv process).