Support seeking in LRIT files
The file-backed LRIT file would support seeking, but the offset
relative to the beginning of the file included the LRIT header,
whereas it should seek to the beginning of the LRIT data segment.
This commit adds an filebuf derivative that includes an offset in
every call to seek such that it looks as if the start of the file is
located at the start of the LRIT data segment.

This change also adds proper seeking support for the membuf.

Both are needed for decompressing EMWIN ZIP files.