Use packet reader in goesproc
This enables goesproc to directly subscribe to the goesrecv packet
publisher instead of manually setting up a pipe with nanocat.
As the nanomsg reader uses a 5 minute buffer (for the HRIT packet
stream), this should fix any packet drops happening due to long
processing times. The --subscribe argument can be used to point
goesproc to the packet stream of interest.