Handle invalid LRIT header in assembler
Some packets appeared on the GOES-15 LRIT stream that had a zero
length LRIT header which prevented lrit::getHeaderMap to terminate.

Fixing this involved a refactor of session_pdu.cc such that it is no
longer constructed with a transport PDU. After this change, all
transport PDUs are appended to the session PDU by calling the append
function. This returns false if the transport PDU could not be
appended to the session PDU (for example if the header is invalid).

As a side effect we can now also finish image session PDUs that didn't
have a final packet (this was not handled before).