Update clock recovery algorithm
First change in this commit is that the PLL gain values are derived
from a single loop bandwidth parameter, according to this article:
http://www.trondeau.com/blog/2011/8/13/control-loop-gain-values.html
This is the same approach used for Costas PLL gain values.

Second change in this commit is that it is now unrolled. The previous
implementation allowed parts to be vectorized (in chunks of 4
samples), but at the cost of stability. I did some tests and found
that the previous implementation could not easily recover from an
omega being much lower than expected. Since no parts of the inner loop
were actually vectorized, the performance penalty of unrolling is
relatively low, at only 15%. This cost is acceptable with the amount
of CPU left to spare for real time demodulation on the Raspberry Pi.

The loop bandwidth is statically initialized to 1e-3 and anecdotal
tests show that it recovers from 1% lower and 1% higher omega values
in a matter of seconds. These are worst case values that were only
useful to test convergence starting from both lower and higher values.