Add tool to compute area intersection
This computes the largest intersection between areas covered by
different products. It spits out a crop value that can be used by
goesproc. The resulting images from different products can then be
used in sequence to get more frequent imagery of a smaller area.

This is largely copied from the previous goesproc version where the
largest intersecting area was computed on the fly. It cannot be done
on the fly in streaming mode, so with the current goesproc version you
need to precompute the cropping coordinates.

This is only relevant for the GOES-N series products because GOES-R
doesn't have different fixed regions (the only fixed region is the
full disk).