Handle multiple regions per instance
This means fewer handlers and a smaller configuration file.

To achieve this the image handlers now key the LRIT files off of both
their region and their channel. Before, it would also key them off of
their image identifier. This should not be needed though, since in
streaming mode these files are processed in order, and never are two
images of the same region and channel sent concurrently.

A side effect of this is that we can now drop incomplete images. If a
new LRIT file is seen with a different image identifier than the ones
the handler already saw, then we know the previous image is incomplete
and drop it.