Fix Himawari-8 area handling again
See c8864547 for the first attempt. That commit fixed handling of
older images but broke handling of newer images. I tested this commit
against both and verified it produces the right output. Unfortunately
the LRIT headers for Himawari-8 images seem to be inconsistent with
the GOES ones and require special handling.