Add goescat utility
In contrast to the packetdump tool in the decoder directory (you had
to know it existed...), goescat allows you to directly subscribe to a
goesrecv packet publisher.

Some more context: to subscribe to a packet publisher when using the
packetdump tool you had to use nanocat, which uses a non-configurable
receive buffer of 128KB. Disk I/O performed by packetdump could take
more than a few seconds if the system is under heavy load, causing the
buffers to back up, and nanocat to drop packets. Since we can spare a
few extra megabytes of memory we should never drop packets in the
receive path. With goescat, it is still possible for buffers to back
up and eventually for packets to be dropped, but only after stalling
longer than 5 minutes.