Run sanity check on session PDU before passing upstream
I saw a crash following a garbage session PDU (that was contained in a
single transport PDU). The contents looked like the middle of an NWS
weather report. With this contents interpreted as a header, it seemed
as if the session PDU would be gigabytes in size, which in turn caused
bad reads. To prevent this from happening, perform a sanity check on
the session PDU size before passing it to the caller.