Make maximum frequency deviation configurable
If you lose signal for a little while, it can happen that the applied
frequency correction tracks far beyond where it should be. This makes
locking back onto the signal if it comes back much harder. This commit
makes the maximum frequency correction configurable, so you can set it
to stay within the bounds that are acceptable for your receiver.

A maximum deviation of 20 KHz should be plenty for both the RTL-SDR
and Airspy units, so this is the default.