Use filename builder of first channel in false color mode
This fixes an issue where running goesproc on the same LRIT files
multiple times can lead to two separate but identical false color
images being generated, due to non-determinism on the order in which
the channels are seen. With this fix, the filename builder is kept
around for both channels, and the one of the first listed channel is
used. Regardless of the order with which the LRIT files are seen, it
will produce the same output filename.