Force setting BUILD_SHARED_LIBS to OFF
Without setting it as a CACHE entry, libaec will try to build shared
libraries after running CMake for the first time. It will reconfigure
to build static libraries upon running CMake for a second time.

I guess this has something to do with the CMake cache but haven't
narrowed down the exact cause.