Work on cross compilation for Raspberry Pi (Raspbian)
It now uses the hosts arm-linux-gnueabihf compiler (GCC version 5) and
the preparation script downloads a larger set of Raspbian libraries
such that goesproc can be compiled as well.

The official Raspberry Pi cross compilation tools ship with GCC 4.9.
Its linker is too old and can't link against the OpenCV libraries that
ship with Raspbian (which are compiled with GCC 6).