Refactor goesproc GOES-R handler
This change attempts too keep the source LRIT files around for as long
as possible contrary to the previous approach where they were turned
into an image as soon as possible. A few things are derived from the
LRIT headers that are convenient to delay to a later stage, such as
the LRIT details and the file name builder. Before this change these
were captured in the Tuple struct. With the pending addition of map
overlays, the navigation header needs to be available as well. Instead
of retrieving it up front and storing it in the Tuple struct, it can
now be retrieved from the right LRIT file when it is needed.