Add map drawer
It is hooked up to the GOES-R and Himawari-8 handlers.

The map overlay can be configured with a "[[handler.map]]" section
that at a minimum defines a "path" field with a path to a GeoJSON
file. The color of the lines can be configured with a "color" field
that contains a hexadecimal color code.