Fix out of bounds issue for bitmap NWS images
This was causing memory corruption related crashes on ARM. If the
dimensions of an image are not a multiple of 8, the code would run a
couple of bytes over the bounds.

My hypothesis is that this was not a problem on 64-bit systems because
allocations are aligned to multiples of 8 anyway, so even though it
would run over a bit, it would do so in memory that isn't used anyway.
On 32-bit systems allocations are aligned to multiples of 4, which has
the risk of running over and clobbering another allocation.