Show packet information when running in TTY
With this change goesproc now logs a single line with packet
information to stdout if stdout is a terminal. If the output is not a
TTY it remains silent and this change has no effect.

Having this feedback means you immediately know whether or not
goesproc is connected to goesrecv and receiving packets. If it isn't
receiving packets the line will show "Waiting for first packet...".
This was ambiguous in the past and you had to wait a while to get
confirmation whether or not packets were flowing by seeing a file get
generated.