Fix for ancillary text header with 'Channel = N/A'
The HRIT stream started to have an entry in the ancillary text LRIT
header equal to 'Channel = N/A'. As goesproc expected this to be an
integer, it caused an exception. In this fix, fall back to using the
product sub ID from the NOAA LRIT header, which appears to be equal to
the channel number.