Try to extract region from ancillary text header first
The region field in the ancillary text header was originally ignored
and the code only used the file name to determine the region
information. This code was specific to CMIP files and didn't work for
non-CMIP files. This change uses the region field in the ancillary
text header and defers to the file name only for CMIP mesoscale files.

Also see #57 for the original issue about non-CMIP files.