Include product in segment key
This is not strictly needed today because all non-CMIP files that
we've seen so far are not segmented. In case they will be at a later
point in time, we include the product name in the key that locates the
set of segments for a new file. If we don't, we'd see segments from
different products be combined.