Fix NWS image timestamp extraction
The pattern that encodes the timestamp for NWS images used to be irregular and
required careful parsing. This pattern was changed some time before November
2020 and became regular. The irregular parser produced invalid timestamps when
applied to filenames with the new pattern. This commit updates the function
that extracts a timestamp from the filename to expect the regular pattern.

This commit also removes the timestamp prefix from the filename if it could be
extracted, allowing users to specify their own timestamp pattern.

This means that a handler with the following definition:
```
filename = "{time:%Y%m%dT%H%M%SZ}_{filename}"
```

No longer produces filenames:
```
20220327T110046Z_20220327110046-hiwind_pac_latest.gif
```

But rather:
```
20220327T110046Z_hiwind_pac_latest.gif
```

Fixes #100.