Strip prefixed v from librtlsdr versions
Some versions of librtlsdr (including versions <= 0.5.4 built directly
from source) have a leading 'v' on their versions as reported by
pkg-config. This causes the version check for bias tees to yield
incorrect results, preventing the use of a bias tee even if the
underlying librtlsdr supports it. Removing the 'v' avoids this.